src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/events-stopped-at-shadow-boundary-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests to ensure that all kinds of events are not stopeed at shadow boundary if created by users.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

None of the events should be stopped at a Shadow boundary if created by users.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
